Fixes bug when replacing an item failed to merge back with the existing item.
The "item" data formats are a little bit different when they are
in different parts of the tree. For example, when in the inventory, an
item has an X and Y field. When in the equipment collection, it doesn't
have those, but has a "attached" boolean.

When an item is generated, it needs to merge back to the collection so
any existing data is retained.