Item generation no longer takes character level into consideration
This has been the source of many questions. Users would try generating an item,
only to be told that it "could not be generated". The reason was their character
level was not high enough. The justification was this allows disambiguation when
there are more than one affixes have the same name.

In retrospect, this is counter-intuitive. The users are using the editor to
generate items. It shouldn't matter what their character levels are. If it can
be found/made, it should work.

The editor has been updated so it always generates the highest quality/level
item it can, unless a level cap is given, which causes it to lower the quality
of the item to match the level cap.