Quest progress now loads on character load
The data is hooked into the character sheet so the "set" and "show" commands can
operate on the data structure unaltered. The quest data is not currently saved
when the character is saved.

The quest data structure doesn't hold sufficient information for it to be useful
for end users. It's questionable how useful reading the quest structure will be.