Refactors how global/db is accessed
Instead of accessing it directly, we're now going through a getter function.
This lets us throw an exception when the editor needs to access the db when the
db has not been loaded.

This also allows us to give the user a more informative error message rather
than just showing a stacktrace that is "near" a failed db lookup.