ArcRecordByteBuffer appears to be working.
Now, there is away to retrieve compressed files inside Arc files without
decompressing the entire file all at once. We're trying to retrieve a small
number of bytes inside the map file that represents the rift gates and shrines.
It would not have been reasonable to try to uncompress the entire 700MB+ file
just to retrieve a handful of bytes.

It would be nice to have an object that looked like a ByteBuffer where we can
decompress arc packed files on the fly since most of the other binary data
reading code expects a ByteBuffer to work with. Unfortunately, this appears to
be quite difficult to do this directly from clojure using Java interop
mechanisms. (It's not clear if it would be any easier or possible at all from
Java directly.)

Anyway, this commit implements something via clojure's defprotocol. This is
still not quite ideal. It would be far nicer to have functions that can take a
byte array and return some kind of primitive, instead of complecting binary data
reading from data interpretation.