Fixes "class add" and "class remove" for certain masteries
We're using UI records to locate the list of masteries the user can see in game.
We then attempt to navigate towards the mastery record. This used to depend on
the class skills being defined in a particular order, where the first skills is
the mastery record. Some masteries defined by "Dawn of Masteries" doesn't seem
to follow this convention.

We're now following all the skill references for two levels until we find a
mastery record.